proyecto/lightning-mcfilm/backend/user_service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
pymysql
python-jose[cryptography]
python-multipart
bcrypt==4.1.3
passlib[bcrypt]==1.7.4